==> code_model/1_Pre-processing_for_Continuous_Targets.ipynb <==
---

## Imports

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler

import pickle
import wget, os
import time
import glob
import pytz

---

## Function Definitions

def summarize(df):
    print(f'{df.shape[0]} rows,   {df.isna().sum().sum()} nulls')
    print(f'\n {df.head(3)}')
    return

---

## Read in Consolidated Data & Examine

with open('../data/consolidated_data.pkl', 'rb') as f:
    df = pickle.load(f)
df.head()

summarize(df)

---

## Add `year`, `month`, `day`, and  `hour` Columns  
(drop the `node`, `date`, and `hr` columns

df.drop(columns=['node', 'date', 'hr_index'], inplace=True)
df.head(3)

df['year']  = df.index.year
df['month'] = df.index.month
df['day']   = df.index.day
df['hour']  = df.index.hour

df.head(3)

df.shape

df.to_csv('../data/pre_processed_data.csv')

with open('../data/pre_processed_df.pkl', 'wb') as f:
    pickle.dump(df, f)

---

## Train-test-split and Scale the Data

train_set_length = int(round((df.shape[0] * 0.75), 0))
test_set_length  = df.shape[0] - train_set_length

print(f'\nTrain: {train_set_length} rows\nTest:  {test_set_length} rows')

train = df.iloc[: train_set_length]
test  = df.iloc[train_set_length: ]

print(f'\nTrain set shape: {train.shape}\nTest set shape:  {test.shape}\n')

train.head(3)

test.head()

ss = StandardScaler()

train_sc = ss.fit_transform(train)
test_sc  = ss.transform(test)

train_sc.shape

test_sc.shape

train_sc_df = pd.DataFrame(train_sc, columns=train.columns, index=train.index)
test_sc_df = pd.DataFrame(test_sc, columns=test.columns, index=test.index)

train_sc_df.shape

test_sc_df.shape

train_sc_df.head(3)

train.to_csv('../data/processed/train.csv')
test.to_csv('../data/processed/test.csv')

with open('../data/processed/train.pkl', 'wb') as f:
    pickle.dump(train, f)
    
with open('../data/processed/test.pkl', 'wb') as f:
    pickle.dump(test, f)

with open('../data/processed/train_sc_df.pkl', 'wb') as f:
    pickle.dump(train_sc_df, f)
    
with open('../data/processed/test_sc_df.pkl', 'wb') as f:
    pickle.dump(test_sc_df, f)
