==> code_model/4_SARIMAX_Model_Cont_Targets_scaled.ipynb <==
---

## Imports

import pandas as pd
import numpy as np

import pickle
import wget, os
import time
import glob
import pytz

# from sklearn.preprocessing import StandardScaler

import statsmodels.api as sm
from statsmodels.tsa.stattools import acf, pacf
from statsmodels.graphics.tsaplots import plot_acf, plot_pacf
from statsmodels.tsa.arima_model import ARIMA
from statsmodels.tsa.seasonal import seasonal_decompose
from statsmodels.tsa.stattools import adfuller
from sklearn.metrics import mean_squared_error
from statsmodels.tsa.statespace.sarimax import SARIMAX


---

## Define Functions

# From Matt Brems lecture

def MSE(true, predicted):
    squared_diff = np.square(true - predicted)
    return np.mean(squared_diff)

# Root Mean Square Error
def RMSE(true, predicted):
    squared_diff = np.square(true - predicted)    
    return np.sqrt(np.mean(squared_diff))

# R-squared, coefficient of determination
def R_squared(true, predicted):
    true      = np.array(true)
    predicted = np.array(predicted)
    sum_squared_diff = sum(np.square(true - predicted))
    variance  = sum(np.square(true - np.mean(true)))
    calc_r2   = 1 - (sum_squared_diff / variance)
    return calc_r2

---

## Load Pickles:  Train/Test Dataframes &  Scaled Arrays

with open('../data/processed/train_sc_df.pkl', 'rb') as f:
    train_sc_df = pickle.load(f)
    
with open('../data/processed/test_sc_df.pkl', 'rb') as f:
    test_sc_df = pickle.load(f)
        
with open('../data/pre_processed_df.pkl', 'rb') as f:
    df = pickle.load(f)

with open('../data/processed/train.pkl', 'rb') as f:
    train = pickle.load(f)
        
with open('../data/processed/test.pkl', 'rb') as f:
    test = pickle.load(f)


exog_col_for_dam = [train.columns[i] for i, col in enumerate(df.columns) if col != 'dam_price_per_mwh']
exog_col_for_hasp = [train.columns[i] for i, col in enumerate(df.columns) if col != 'hasp_price_per_mwh']

X_train_sc_dam = train_sc_df[exog_col_for_dam]
X_test_sc_dam  = test_sc_df[exog_col_for_dam]
y_train_dam    = train['dam_price_per_mwh']
y_test_dam     = test['dam_price_per_mwh']

X_train_sc_hasp = train_sc_df[exog_col_for_hasp]
X_test_sc_hasp  = test_sc_df[exog_col_for_hasp]
y_train_hasp    = train['hasp_price_per_mwh']
y_test_hasp     = test['hasp_price_per_mwh']

y_train_dam.shape

X_train_sc_dam.shape

---

## Gridsearch to find best MSE params for `P`, `D`, `Q`, and `S`

(for `DAM` and `HASP` prices as endogenous variables,  
 use the ARIMA parameters for `p`, `d`, & `q` found in previous ARIMA gridsearch...  
 `DAM: 4, 0, 6`  and  `HASP: 6, 0 ,6`)

> NOTE: Because of the amount of time each SARIMAX model takes to fit, there was simply not enough time to do these grid searches, although the code is correct and ready to run

# Matt Brem's grid search loop
# Starting MSE and (P, D, Q).

# mse = 99 * (10 ** 16)
# final_P = 0
# final_D = 0
# final_Q = 0
# final_S = 0

# for P in range(3):
#     for Q in range(3):
#         for D in range(3):
#             for S in range(0,24,8):
#                 try:
#                     # Instantiate SARIMA model.
#                     sarima = SARIMAX(endog = y_train_dam,
#                                      order = (4, 0, 6),              # (p, d, q)
#                                      seasonal_order = (P, D, Q, S),  # (P, D, Q, S)
#                                      exog = X_train_sc_dam) 

#                     # Fit SARIMA model.
#                     model = sarima.fit()

#                     # Generate predictions based on training set.
#                     preds = model.predict()

#                     # Evaluate predictions.
#                     print(f'MSE for (4,0,6) x ({P},{D},{Q},{S}) ... {mean_squared_error(train["dam_price_per_mwh"], preds)}')

#                     # Save for final report.
#                     if mse > mean_squared_error(train['dam_price_per_mwh'], preds):
#                         mse = mean_squared_error(train['dam_price_per_mwh'], preds)
#                         final_P = P
#                         final_D = D
#                         final_Q = Q
#                         final_S = S

#                 except:
#                     pass

# print(f'Our model that minimizes MSE on the training data is the SARIMA(4,0,6) x ({final_P},{final_D},{final_Q},{final_S})')
# print(f'This model MSE = {mse}')

# Matt Brem's grid search loop
# Starting MSE and (P, D, Q).

# mse = 99 * (10 ** 16)
# final_P = 0
# final_D = 0
# final_Q = 0
# final_S = 0

# for P in range(2):
#     for Q in range(2):
#         for D in range(2):
#             for S in range(0,25,12):
#                 try:
#                     # Instantiate SARIMA model.
#                     sarima = SARIMAX(endog = y_train_hasp],
#                                      order = (6, 0, 6),              # (p, d, q)
#                                      seasonal_order = (P, D, Q, S),  # (P, D, Q, S)
#                                      exog = X_train_sc_hasp) 

#                     # Fit SARIMA model.
#                     model = sarima.fit()

#                     # Generate predictions based on training set.
#                     preds = model.predict()

#                     # Evaluate predictions.
#                     print(f'MSE for (6,0,6) x ({P},{D},{Q},{S}) ... {mean_squared_error(train["hasp_price_per_mwh"], preds)}')

#                     # Save for final report.
#                     if mse > mean_squared_error(train['hasp_price_per_mwh'], preds):
#                         mse = mean_squared_error(train['hasp_price_per_mwh'], preds)
#                         final_P = P
#                         final_D = D
#                         final_Q = Q
#                         final_S = S

#                 except:
#                     pass

# print(f'Our model that minimizes MSE on the training data is the SARIMA(6,0,6) x ({final_P},{final_D},{final_Q},{final_S})')
# print(f'This model MSE = {mse}')

## Instantiate and fit the models with best params

P = 0
D = 1
Q = 0
S = 24

# Instantiate SARIMA model.
dam_sarimax01024 = SARIMAX(endog = y_train_dam,
                 order = (4, 0, 6),              # (p, d, q)
                 seasonal_order = (P, D, Q, S),  # (P, D, Q, S)
                 exog = X_train_sc_dam) 

# Fit SARIMA model.
dam_sarimax01024_model = dam_sarimax01024.fit()

# Generate predictions based on training set.
dam_sarimax01024_preds = dam_sarimax01024_model.predict()

# Evaluate predictions.
print(f'MSE for (4,0,6) x ({P},{D},{Q},{S}) ... {mean_squared_error(train["dam_price_per_mwh"], dam_sarimax01024_preds):.2f}')
 


with open('../fitted_models/dam_sarimax01024_model.pkl', 'wb') as f:
    pickle.dump(dam_sarimax01024_model, f)
    
with open('../data/predictions/dam_sarimax01024_preds.pkl', 'wb') as f:
    pickle.dump(dam_sarimax01024_preds, f)

P = 0
D = 1
Q = 0
S = 24

# Instantiate SARIMA model.
hasp_sarimax01024 = SARIMAX(endog = y_train_hasp,
                 order = (6, 0, 6),              # (p, d, q)
                 seasonal_order = (P, D, Q, S),  # (P, D, Q, S)
                 exog = X_train_sc_hasp) 

# Fit SARIMA model.
hasp_sarimax01024_model = hasp_sarimax01024.fit()

# Generate predictions based on training set.
hasp_sarimax01024_preds = hasp_sarimax01024_model.predict()

# Evaluate predictions.
print(f'MSE for (6,0,6) x ({P},{D},{Q},{S}) ... {mean_squared_error(train["hasp_price_per_mwh"], hasp_sarimax01024_preds):.2f}')
 


with open('../fitted_models/hasp_sarimax01024_model.pkl', 'wb') as f:
    pickle.dump(hasp_sarimax01024_model, f)
    
with open('../data/predictions/hasp_sarimax01024_preds.pkl', 'wb') as f:
    pickle.dump(hasp_sarimax01024_preds, f)
