==> code_model/5_Model_Evaluation.ipynb <==
---

## Imports

import pandas as pd
import numpy as np

import pickle
import time
import glob
import pytz

import matplotlib.pyplot as plt
import seaborn as sns
from pandas.plotting import register_matplotlib_converters
 
register_matplotlib_converters()
%matplotlib inline
sns.set_style('whitegrid')

---

## Function Definitions

# From Matt Brems lecture

def MSE(true, predicted):
    squared_diff = np.square(true - predicted)
    return np.mean(squared_diff)

# Root Mean Square Error
def RMSE(true, predicted):
    squared_diff = np.square(true - predicted)    
    return np.sqrt(np.mean(squared_diff))

# R-squared, coefficient of determination
def R_squared(true, predicted):
    true      = np.array(true)
    predicted = np.array(predicted)
    sum_squared_diff = sum(np.square(true - predicted))
    variance  = sum(np.square(true - np.mean(true)))
    calc_r2   = 1 - (sum_squared_diff / variance)
    return calc_r2

---

## Unpickle Predictions Variables, Dataframes  

\* Please note:  graphs and $r^2$ scores for dam and hasp rnn model runs are contained in those individual notebooks, numbered `7a`, `7b`, & `7c` for day-ahead market runs, and `8a`, `8b`, & `8c` for hour-ahead market runs, respectively.

ARIMA and SARIMAX:
 - dam_arima406_preds.pkl  
 - hasp_arima606_preds.pkl  
  
 - dam_sarimax01024_preds.pkl  
 - hasp_sarimax01024_preds.pkl   

with open('../data/pre_processed_df.pkl', 'rb') as f:
    df = pickle.load(f)

with open('../data/processed/train.pkl', 'rb') as f:
    train = pickle.load(f)
    
with open('../data/processed/test.pkl', 'rb') as f:
    test = pickle.load(f)

with open('../data/predictions/*.pkl', 'rb') as f:
    * = pickle.load(f)

with open('../data/predictions/dam_arima406_preds.pkl', 'rb') as f:
    dam_arima406_preds = pickle.load(f)
    
with open('../data/predictions/hasp_arima606_preds.pkl', 'rb') as f:
    hasp_arima606_preds = pickle.load(f)

with open('../data/predictions/dam_sarimax01024_preds.pkl', 'rb') as f:
    dam_sarimax01024_preds = pickle.load(f)
    
with open('../data/predictions/hasp_sarimax01024_preds.pkl', 'rb') as f:
    hasp_sarimax01024_preds = pickle.load(f)
    

---

## ARIMA Predictions, MSE, and $R^2$ Scores

print(train.shape[0])
print(test.shape[0])

mse_dam_arima406_train = MSE(df["dam_price_per_mwh"][:train.shape[0]], np.array(dam_arima406_preds)[:train.shape[0]])


r2_dam_arima406_train = R_squared(df["dam_price_per_mwh"][:train.shape[0]], np.array(dam_arima406_preds)[:train.shape[0]])


print(f'MSE for dam_arima_406_train ... {mse_dam_arima406_train:.2f}')

print()
print(f'R-sq for dam_arima_406_train ... {r2_dam_arima406_train:.4f}')




mse_hasp_arima606_train = MSE(df["hasp_price_per_mwh"][:train.shape[0]], np.array(hasp_arima606_preds)[:train.shape[0]])

r2_hasp_arima606_train = R_squared(df["hasp_price_per_mwh"][:train.shape[0]], np.array(hasp_arima606_preds)[:train.shape[0]])


print(f'MSE for hasp_arima_606_train ... {mse_hasp_arima606_train:.2f}')
print()
print(f'R-sq for hasp_arima_606_train ... {r2_hasp_arima606_train:.4f}')




---

## SARIMAX Predictions, MSE, and $R^2$ Scores

mse_dam_sarimax01024_train = MSE(train["dam_price_per_mwh"], np.array(dam_sarimax01024_preds))
r2_dam_sarimax01024_train = R_squared(train["dam_price_per_mwh"], np.array(dam_sarimax01024_preds))

mse_hasp_sarimax01024_train = MSE(train["hasp_price_per_mwh"], np.array(hasp_sarimax01024_preds))
r2_hasp_sarimax01024_train = R_squared(train["hasp_price_per_mwh"], np.array(hasp_sarimax01024_preds))


print(f'MSE for dam_sarimax01024_train ... {mse_dam_sarimax01024_train:.2f}')
print(f'R-sq for dam_sarimax01024_train ... {r2_dam_sarimax01024_train:.4f}')
print()
print(f'MSE for hasp_sarimax01024_train ... {mse_hasp_sarimax01024_train:.2f}')
print(f'R-sq for hasp_sarimax01024_train ... {r2_hasp_sarimax01024_train:.4f}')
